+++ CHANGELOG.md
@@ -34,6 +34,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 *   Fixed current range "zoom out" when output switch is "off".
 *   Added on/off switch #84.
 *   Added horizontal markers to the waveform widget #37.
+*   Added statistics on/off for all vertical markers #12.
 
 
 ## 0.8.16